Refactor client connection details to new module.
